backend/requirements.txt
flask
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
flask-bcrypt
flask-pyjwt
psycopg2